requirements.txt
numpy
opencv-python
pycryptodome
matplotlib
